remove circular dependencies in GFF parsing code
