+++ js/feature/gff/gffHelper.js
@@ -1,32 +1,5 @@
-/*
- * The MIT License (MIT)
- *
- * Copyright (c) 2021 University of California San Diego
- * Author: Jim Robinson
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- */
-
 import {isExon, isTranscript, isTranscriptPart} from "./so.js"
 import {GFFFeature, GFFTranscript} from "./gffFeature.js"
-import {parseAttributeString} from "./gff.js"
 
 const gffNameFields = ["Name", "gene_name", "gene", "gene_id", "alias", "locus", "name"]
 
@@ -229,16 +202,15 @@ class GFFHelper {
     nameFeatures(features) {
         // Find name (label) property
         for (let f of features) {
-            if (f.attributeString) {
-                const delim = ('gff3' === this.format) ? '=' : ' '
-                const attributes = parseAttributeString(f.attributeString, delim)
-                const attributesMap = new Map(attributes)
+            if(typeof f.getAttributeValue === 'function') {
+
                 if (this.nameField) {
-                    f.name = attributesMap.get(this.nameField)
+                    f.name = f.getAttributeValue(this.nameField)
                 } else {
                     for (let nameField of gffNameFields) {
-                        if (attributesMap.has(nameField)) {
-                            f.name = attributesMap.get(nameField)
+                        const v = f.getAttributeValue(nameField)
+                        if (v) {
+                            f.name = v
                             break
                         }
                     }
@@ -250,3 +222,4 @@ class GFFHelper {
 
 
 export default GFFHelper
+
